backend/requirements.txt
bcrypt~=4.3.0
pytest~=8.3.5
cryptography~=44.0.2
fastapi~=0.115.12
python-jose~=3.4.0
python-multipart~=0.0.20
uvicorn~=0.34.2
uvloop~=0.21.0
aiofiles~=24.1.0
passlib~=1.7.4
beautifulsoup4~=4.13.4
soupsieve~=2.7
typing_extensions~=4.13.2
h11~=0.14.0
pip~=23.2.1
google~=3.0.0
rsa~=4.9.1
pyasn1~=0.4.8
sniffio~=1.3.1
click~=8.1.8
httpcore~=1.0.8
idna~=3.10
httpx~=0.28.1
pyasn1-modules==0.2.8
cachetools~=5.5.2
certifi~=2025.1.31
urllib3~=2.4.0
requests~=2.32.3
cffi~=1.17.1
pydantic~=2.11.3
websockets~=15.0.1
anyio~=4.9.0
pydantic_core~=2.33.1
annotated-types~=0.7.0
charset-normalizer~=3.4.1
pymongo~=4.12.0
dotenv~=0.9.9
python-dotenv~=1.1.0
dnspython~=2.7.0
ecdsa~=0.19.1
pytz~=2025.2
six~=1.17.0
python-dateutil~=2.9.0.post0
starlette~=0.46.2
psycopg2-binary~=2.9.10
pycparser~=2.22
packaging>=24.2
bleach~=6.2.0
attrs~=25.3.0
tornado~=6.5
Jinja2~=3.1.6
pyzmq~=26.4.0
pexpect~=4.9.0
parso~=0.8.4
jedi~=0.19.2
Pygments~=2.19.1
yarg~=0.1.9
webencodings~=0.5.1
tinycss2~=1.4.0
ipython~=8.12.3
MarkupSafe~=3.0.2
plotly~=6.1.1
narwhals~=1.40.0
traitlets~=5.14.3
stack-data~=0.6.3
decorator~=5.2.1
matplotlib-inline~=0.1.7
backcall~=0.2.0
pickleshare~=0.7.5
jupyter_client~=8.6.3
prompt_toolkit~=3.0.51
wcwidth~=0.2.13
ptyprocess~=0.7.0
docopt~=0.6.2
nbformat~=5.10.4
jupyter_core~=5.7.2
nbclient~=0.10.2
fastjsonschema~=2.21.1
jsonschema~=4.23.0
asttokens~=3.0.0
executing~=2.2.0
nbconvert~=7.16.6
pandocfilters~=1.5.1
defusedxml~=0.7.1
mistune~=3.1.3
jupyterlab_pygments~=0.3.0
pure_eval~=0.2.3
referencing~=0.36.2
jsonschema-specifications~=2025.4.1
platformdirs~=4.3.8
pillow>=11.2.1
protobuf>=5.29.5
setuptools>=80.8.0
grpcio>=1.71.0
PyYAML>=6.0.2
multidict>=6.4.4
propcache>=0.3.1
zstandard>=0.23.0
proto-plus>=1.26.1
regex>=2024.11.6
threadpoolctl>=3.6.0
mpmath>=1.3.0
sympy>=1.14.0
filelock>=3.18.0
tqdm>=4.67.1
networkx>=3.4.2
fsspec>=2025.5.0
aiohttp>=3.11.18
yarl>=1.20.0
grpcio-status>=1.71.0
joblib>=1.5.0
orjson>=3.10.18
SQLAlchemy>=2.0.41
aiosignal>=1.3.2
frozenlist>=1.6.0
aiohappyeyeballs>=2.6.1
filetype>=1.2.0
greenlet>=3.2.2
tenacity>=9.1.2
requests-toolbelt>=1.0.0
jsonpointer>=3.0.0
tokenizers>=0.21.1
marshmallow>=3.26.1
jsonpatch>=1.33
typing-inspect>=0.9.0
dataclasses-json>=0.6.7
pydantic-settings>=2.9.1
mypy_extensions>=1.1.0
httpx-sse>=0.4.0
langchain-google-genai>=2.1.5
langchain-core>=0.3.62
langchain>=0.3.25